Move the global variables initialization code to one function
Previously, the code that was initialiaing the global variables and the
code that was rendering the page were in different window-load handlers,
so execution order was not guaranteed and this could cause race
condition. Now the order is fixed and there should be no surprises :)
