+++ templates/welcome.html
@@ -16,6 +16,7 @@
                 <script type="text/javascript" src="{{ url_for('static', filename='js/render_calendar.js') }}"></script>
                 <script type="text/javascript" src="{{ url_for('static', filename='js/render_remove_datapoint.js') }}"></script>
                 <script type="text/javascript" src="{{ url_for('static', filename='js/global_vars.js') }}"></script>
+                <script type="text/javascript" src="{{ url_for('static', filename='js/init.js') }}"></script>
                 <form id="delete_habit">
                 </form>
